desktop/requirements.txt
librosa
scikit-learn
matplotlib
seaborn
umap-learn
